Create a cookie utilities file; create a PUT /user endpoint
