+++ services/admin-api/src/routes/user.js
@@ -1,19 +1,10 @@
 const { rootDb } = require("@your-analytics/faunadb");
 const { Magic } = require("@magic-sdk/admin");
 const express = require("express");
-const jwt = require("jsonwebtoken");
+const { cookieConfig, signJwtAndSetCookie } = require("../utils/cookies");
 
 const magic = new Magic(process.env.MAGIC_SECRET_KEY);
 const router = express.Router();
-
-const cookieConfig = {
-  path: "/",
-  maxAge: 1000 * 60 * 60 * 24 * 7,
-  httpOnly: true,
-  sameSite: true,
-  signed: true,
-  secure: true,
-};
 
 module.exports = (authenticateMagic, authenticateJwtCookieCombo) => {
   router.get("/", authenticateJwtCookieCombo, async (req, res) => {
@@ -24,17 +15,32 @@ module.exports = (authenticateMagic, authenticateJwtCookieCombo) => {
     return res.status(404).end();
   });
 
+  router.put("/", authenticateJwtCookieCombo, async (req, res) => {
+    try {
+      await rootDb.users.update(req.user.issuer, req.body);
+      const dbUser = await rootDb.users.find(req.user.issuer);
+
+      try {
+        await signJwtAndSetCookie(res, dbUser);
+        return res.status(200).json(req.user).end();
+      } catch (error) {
+        return res.status(401).end("Something went wrong.");
+      }
+    } catch (error) {
+      console.error(error);
+      return res.status(500).end();
+    }
+  });
+
   router.post("/login", authenticateMagic, async (req, res) => {
     if (req.user) {
       const dbUser = await rootDb.users.find(req.user.issuer);
-      jwt.sign({ user: dbUser }, process.env.JWT_SECRET, (error, token) => {
-        if (error) {
-          return res.status(401).end("Could not log user in.");
-        }
-
-        res.cookie("jwt", token, cookieConfig);
+      try {
+        await signJwtAndSetCookie(res, dbUser);
         return res.status(200).json(req.user).end();
-      });
+      } catch (error) {
+        return res.status(401).end("Could not log user in.");
+      }
     } else {
       return res.status(401).end("Could not log user in.");
     }
